chore: re-organize registering of plugins and routes
