+++ packages/dashboard-backend/src/plugins/staticServer.ts
