+++ packages/dashboard-backend/src/plugins/swagger.ts
