+++ packages/dashboard-backend/src/plugins/webSocket.ts
@@ -1,0 +1,18 @@
+/*
+ * Copyright (c) 2018-2021 Red Hat, Inc.
+ * This program and the accompanying materials are made
+ * available under the terms of the Eclipse Public License 2.0
+ * which is available at https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *
+ * Contributors:
+ *   Red Hat, Inc. - initial API and implementation
+ */
+
+import fastifyWebsocket from '@fastify/websocket';
+import { FastifyInstance } from 'fastify';
+
+export async function registerWebSocket(server: FastifyInstance): Promise<void> {
+  await server.register(fastifyWebsocket);
+}
